Add blob migration utility function.
