+++ datajoint/__init__.py
@@ -41,5 +41,6 @@ from .fetch import key
 from .attribute_adapter import AttributeAdapter
 from . import errors
 from .errors import DataJointError
+from .migrate import migrate_dj11_external_blob_storage_to_dj12
 
 ERD = Di = Diagram   # Aliases for Diagram
